Define a User type for holding user provisioning details
The list of supplementary groups isn't portable and needs to be exposed
in the public API somehow. Rather than expand the Provision API to
include it directly, add a User type that holds the username, password,
ssh keys, and supplementary group information and accept that in the
provision interface.

To start with I left the default group list at "wheel" since it covers
most distributions and if the user is in wheel they can add themselves
to the other groups if they need it later.